Add an .eslintrc file to share coding conventions
These rules are potentially too strict, so feel free to propose
relaxations. These rules are for eslint version 2 (currently 2.4.0), and
they include all of the node-style-guide
(https://github.com/felixge/node-style-guide) rules.

Upgrading to ES2015: If we ever advance the project to a version of node
that supports ES2015, we want to add new rules for the new syntax. In
addition, the no-inner-declarations rule will be out of date.