Fix stats master test
The test `should count total incidents` was failing occasionally because
the test would run before the statsMaster updated.

To reproduce the error this fixes, run the `should count total
incidents` test alone (by calling `it.only`).

This fix causes the test to wait until statsMaster stops emitting
events (for 100ms). This is an imperfect fix, as under high load the
`stats` events could be delayed arbitrarily.