Bugfix for change password button
Summary: Subtly poor coding practices from long ago became a bug when
internationalization was added.

To reproduce the bug this fixes, go to `/users`, click the pencil by
your user, and click the Change Password button. This button wasn't
doing anything.

The cause was that adding the `translate` directive to the element
creates a new child scope. When the `ng-click` directive tried to update
the `showPassword` model it was only changing on the child scope. Best
practice is for `$scope` to always hold a reference to an object that
holds the primitives that make up the model. Then the object can be
shared between parent and child scopes. See
https://github.com/angular/angular.js/wiki/Understanding-Scopes

I read through the rest of the internationalization changes and it
seems this is the only place this comes up.