Run protractor tests on aggie-test database
`npm run testrun` is now the analogue of `npm start` to be used during
protractor tests: this connects to the `aggie-test` databse. The
protractor tests delete everything from this database before every test,
so it's important we're not running on the `aggie` database, or it could
be easy to wipe out everything.