Pass query back to /reports when leaving batch
If you leave a batch through clicking Cancel or Mark all and done or by
clicking Mark all and grab another when there are no more, you are sent
back to `/reports`. With this change, that reports page still has your
filter applied.

The `page` parameter is lost because batch doesn't care about it. It's
unclear to me which page you should expect to return to (the one you
left from, or the first page?).