Grab batch with keyword search
`keywords` used to be a special case of the query filter. With mongo 2.6
(which was from April 2014!) we now have `$text` with which it behaves
like any other part of the query.

N.b. I've removed some logic that limits keyword searches to 100
reports. I couldn't figure out why it was there, so I assume it was in
order to reduce load. Since `$text` is builtin to mongo, I'm hoping that
it is fast enough that we don't have to add this limit. On my machine,
a keyword search that checks 11k reports and returns 10k of them has no
noticeable delay.

Unfortunately, we still depend on `mongoose-text-search` since
`Report.textSearch` is used in `trend-queryer.js`.