Fix editing incidents
Now the backend expects to receive incident objects (creating or
updating) that have a tags field which is an array. The backend
continues to expect the tags field of an incident query to be a string.
This distinction is because incidents are passed in request bodies,
while incident queries are passed in query strings.