Added Update Tags Feature
However the changes made to tags are not reflected until refresh. This will be fixed in a refactor of code to match the way incidents are updated and added. We must also consider the case when tags are deleted. They must also be deleted on every report they are on.