Tag Page, Group Report Tag Add, Individual Report Tag Add (#361)
* Revert "Revert "test: SMTC model""

This reverts commit 5a338b3259f5b119ab08e31d63832a21a8e04cbf.

* Revert "Revert "test: API endpoints for SMTC tags""

This reverts commit 56636b92af0db891f1a03396ee0b6b9c57e017c2.

* Revert "Revert "feat: expose API endpoints for SMTC tags""

This reverts commit be9a6606b400f9e302936934e7cae6f147ee975e.

* Revert "Revert "feat: add SMTCTag model for report-level tagging""

This reverts commit 2145c3b05be4117d8b1c1f9f8d63a0812ace9525.

* Revert "Revert "test: add SMTC model to general db setup & teardown routines""

This reverts commit e752a156a4c57e9325b8b325dc4faa776fd31907.

* Bringing back Tags

Also adding Reports/Last min instead of Reports/min

* Add service to fetch, add and delete SMTC tags

* Added tagify stylesheets and removed select2 libraries

We don't actually use select2 for anything so I have removed the library.

* Add frontend updates for SMTC Tag operations

* Fix typo

* Remove Socket Listeners so we don't create leaks

* Use singular SMTCTag for consistency with backend

* Added Initial Route to Tags Page

* fix: correct schema reference typo

* Adding Tag Page, Modal, Controllers

* Removed Old Test SMTC Language

* Added Tag Pages and Modals

* Fix fronted tag display to use socket.io

* Move to indiana

* Got Descriptioin and Color Inputs Working

TODO: Get Creator and Date Created Working.

* Lighten the gray color; use for background too

* Also remove gray for read reports

* Added username to tag table and Updated Alert Messages

Also removed date and fixed response codes.

* Added Update Tags Feature

However the changes made to tags are not reflected until refresh. This will be fixed in a refactor of code to match the way incidents are updated and added. We must also consider the case when tags are deleted. They must also be deleted on every report they are on.

* Added coloring to Report Page Tags and Edited Tags Modal to remove descriptions

TODO: I need to add colors to Tags Modal

* Changed Colors and made grey colors schemes match Facebook

* Removed Tabs Column from Tag Modal

This is redundant as the tags will be shown in the add & remove tags bar.

* Added Report API endpoint for SMTCTags

This updates the API to be able to add multiple Tags with one API request and this also matches the current method of adding and removing Incidents.

* Re ordering the Alert Bar

* Clear Tags from Selected Functionality

Added ability to clear tags from selected.

* Removing Secondary Color from Search Bars

This coloring isn't actually useful and would be better utilized in tags.

* Finalized Reports

Added measures to prevent duplicating tags and removing tags that aren't there. Also realized that tag.toString was necessary because of the object type in the schema file.

* Added Color Tags to Report Table and Adjusted Models

Changed the modals so that you can see information when adding incidents or tags. Have not added smtcTag functionally for individual reports on the Index page or the show page.

* Tags introduced to Add Tag to Report Modal

The functionality to actually change tags is still TODO. The next commit with include this.

* Individual Tag Editing for Reports

There are remaining TODOs but tags now can be added by report (similiar to incidents).

* Cursor: pointer the Edit and Remove Icons on Tag and Incident Page

The cursor didn't turn into a pointer when hovering over edit and delete icon.

* Fixed Messaging, Removed Debug code and Unnecessary code, and Updated Locale Jsons

I changed messaging from "sorry that report couldn't be saved for some reason" to "sorry that report couldn't be saved." I also added messages for updated tags and removed a bunch of debugging code I used and forgot to remove.

* Made the smtcTags toggle-controlled

Originally the tags were controlled using an add/remove system. This was clearly confusing to a certain extent. I have changed the system. I also added a bunch of comments to the index.html controller in the reports.

* Clicking a Tags on the Report Table Removes It

In addition to the toggle setting I have added a way of removing individual tags on each individual report by clicking a tag in the Tag column of the reports table. This allows for more fine tuned editing of tags without opening the tags modal.

* Tags are now editable from the Report Show Page

Removed a lot of console.logs and corrected the messaging for when no tags are made yet.

* Moved Spectrum Script inside AngularJS Controller

Removed Script tag as it's not good code.

* Tag validation Clarifications

Made it clear what fields were required to create a tag and also created a color palette that imitates trello.

* Fix Updates for Tags

Tags now update without refreshing on edit and creation.

* Responsive table on SMTCtag modal

The table on the modal is too small sometimes and leaks out of the modal. Using a responsive table fixes this problem.

* One last console.log to remove

* Form Validation Errors Show when Violated. Colorpicker starts with no color or edited tag color.

This fixes the validation errors from showing all the time. Color Picker now starts with color or edited tag color.

* Move to Laptop

* Fixed incidents page, bugs surrounding editing tags, custom colors not working

I think this is the last commit necessary before PR is merged.

* Change to Using Aggie Docs to include SMTCTags

* Update Batch Page Tag Sidebar to Match Report Page

* Update Toggle Tag on Batch to Match Report Page

* Optimize new images

Co-authored-by: Min Htat Kyaw <minhtatdx00@gmail.com>
Co-authored-by: Max Karpawich <maxkarpawich@gmail.com>
Co-authored-by: Kevin Cooper <cooperka92@gmail.com>