Make clicking the "___ new reports" flash refresh the page (#413)
This is a rather rudimentary fix but prevents end users from breaking the system unintentionally. I left a TODO in there to make the original method work but usually this is not outside the realm of how systems behave. When there are new changes to the reports, the page will request you refresh the page.