Incident Search Updates (#439)
* Incident Search Updates

Added the ability to search for a specific incident ID, and to filter by author with a dropdown.

* package lock

package lock is showing up as changed locally for some reason. Maybe this is the issue?

* working version

working version of incident search, but the id searched will return an id that shows that id + 1. Fixing this would either change old IDs shown to the user, or break the feature. Trying to find another answer.

* Bugfix

Fixed the issue with idnum either being a vlue off from the number shown to the user, or cause incidents not to play unless their id was explicitly searched.